# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bun Hono Server API",
    "version": "1.0.0",
    "description": "Generated locally via static analyzer"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/users/me": {
      "post": {
        "summary": "更新用户信息",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data",
                    "timestamp"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 20
                  },
                  "nickname": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 30
                  },
                  "avatar": {
                    "type": "string",
                    "format": "uri"
                  },
                  "phone": {
                    "type": "string",
                    "pattern": "^1[3-9]\\d{9}$"
                  },
                  "gender": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 2
                  },
                  "birthday": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "bio": {
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除账户",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data",
                    "timestamp"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/password": {
      "post": {
        "summary": "修改密码",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data",
                    "timestamp"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "minLength": 1
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 50
                  }
                },
                "required": [
                  "oldPassword",
                  "newPassword"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  }
}